Update import paths and references to renamed folders/files
Update imports, exports, and markdown links to reflect the
kebab-case to snake_case renames from the previous commit.
