+++ src/platform/plugins/shared/workflows_execution_engine/server/step/timeout_zone_step/index.ts
@@ -7,7 +7,7 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-export { EnterWorkflowTimeoutZoneNodeImpl } from './workflow-level/enter_workflow_timeout_zone_node_impl';
-export { ExitWorkflowTimeoutZoneNodeImpl } from './workflow-level/exit_workflow_timeout_zone_node_impl';
-export { EnterStepTimeoutZoneNodeImpl } from './step-level/enter_step_timeout_zone_node_impl';
-export { ExitStepTimeoutZoneNodeImpl } from './step-level/exit_step_timeout_zone_node_impl';
+export { EnterWorkflowTimeoutZoneNodeImpl } from './workflow_level/enter_workflow_timeout_zone_node_impl';
+export { ExitWorkflowTimeoutZoneNodeImpl } from './workflow_level/exit_workflow_timeout_zone_node_impl';
+export { EnterStepTimeoutZoneNodeImpl } from './step_level/enter_step_timeout_zone_node_impl';
+export { ExitStepTimeoutZoneNodeImpl } from './step_level/exit_step_timeout_zone_node_impl';
