+++ src/platform/plugins/shared/workflows_execution_engine/server/step/wait_step/README.md
@@ -148,7 +148,7 @@ steps:
 
 **Threshold:** `SHORT_DURATION_THRESHOLD = 5000ms` (5 seconds)
 
-**Duration Parsing:** [`parse-duration.ts`](../../utils/parse-duration/parse-duration.ts)
+**Duration Parsing:** [`parse_duration.ts`](../../utils/parse_duration/parse_duration.ts)
 
 **Task Manager:** Long waits use `workflowTaskManager.scheduleResumeTask()` to schedule resume
