+++ src/platform/plugins/shared/workflows_execution_engine/server/utils/index.ts
@@ -7,7 +7,7 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-export { parseDuration } from './parse-duration/parse-duration';
+export { parseDuration } from './parse_duration/parse_duration';
 export { buildStepExecutionId } from './build_step_execution_id/build_step_execution_id';
 export { stringifyStackFrames } from './stringify_stack_frames';
 export { getKibanaUrl, buildWorkflowExecutionUrl } from './get_kibana_url';
